DEV-9 move PacketParser to domain
