+++ src/DataTransfer/application/data_transfer.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/HyperloopUPV-H8/Backend-H8/DataTransfer/domain"
 	"github.com/HyperloopUPV-H8/Backend-H8/Shared/ExcelParser/application/interfaces"
-	packetAdapter "github.com/HyperloopUPV-H8/Backend-H8/Shared/PacketAdapter/application/interfaces"
+	packetAdapter "github.com/HyperloopUPV-H8/Backend-H8/Shared/PacketAdapter/domain/interfaces"
 
 	"github.com/davecgh/go-spew/spew"
 )
