--- src/Shared/PacketAdapter/application/interfaces/i_measurement.go
@@ -1,6 +1,0 @@
-package interfaces
-
-type Measurement interface {
-	Name() string
-	ValueType() string
-}
