--- src/Shared/PacketAdapter/application/interfaces/i_transport_controller.go
@@ -1,9 +1,0 @@
-package interfaces
-
-type TransportController interface {
-	ReceiveData() []byte
-	ReceiveMessages() [][]byte
-	Send(string, []byte)
-	AliveConnections() []string
-	Close()
-}
