+++ src/Shared/PacketAdapter/domain/interfaces/i_packet.go
